move files out of redesign into src
