+++ src/onemod/__init__.py
@@ -1,0 +1,3 @@
+from onemod.pipeline import Pipeline
+
+__all__ = [Pipeline]
