+++ src/onemod/config/__init__.py
@@ -1,0 +1,8 @@
+from onemod.config.base import (
+    StageConfig,
+    GroupedConfig,
+    CrossedConfig,
+    ModelConfig,
+)
+
+__all__ = [StageConfig, GroupedConfig, CrossedConfig, ModelConfig]
