+++ src/onemod/config/data_config/__init__.py
@@ -1,0 +1,5 @@
+from onemod.redesign.config.data_config.preprocessing_config import (
+    PreprocessingStage,
+)
+
+__all__ = [PreprocessingStage]
