+++ src/onemod/config/data_config/preprocessing_config.py
@@ -1,0 +1,11 @@
+"""Preprocessing stage settings."""
+
+from pathlib import Path
+
+from onemod.redesign.config import StageConfig
+
+
+class PreprocessingConfig(StageConfig):
+    """Preprocessing stage settings."""
+
+    data: Path
