+++ src/onemod/config/model_config/__init__.py
@@ -1,0 +1,5 @@
+from onemod.config.model_config.kreg_config import KregConfig
+from onemod.config.model_config.rover_config import RoverConfig
+from onemod.config.model_config.spxmod_config import SpxmodConfig
+
+__all__ = [KregConfig, RoverConfig, SpxmodConfig]
