+++ src/onemod/config/model_config/kreg_config.py
@@ -1,0 +1,11 @@
+"""Kreg stage settings."""
+
+from pathlib import Path
+
+from onemod.config import ModelConfig
+
+
+class KregConfig(ModelConfig):
+    """Kreg stage settings."""
+
+    offset: Path | None = None
