+++ src/onemod/config/model_config/rover_config.py
@@ -1,0 +1,10 @@
+"""Rover stage settings."""
+
+from onemod.config import ModelConfig
+
+
+class RoverConfig(ModelConfig):
+    """Rover stage settings."""
+
+    cov_exploring: set[str]
+    cov_fixed: set[str] = {"intercept"}
