+++ src/onemod/stage/__init__.py
@@ -1,0 +1,3 @@
+from onemod.stage.base import Stage, GroupedStage, CrossedStage, ModelStage
+
+__all__ = [Stage, GroupedStage, CrossedStage, ModelStage]
