+++ src/onemod/stage/data_stage/__init__.py
@@ -1,0 +1,3 @@
+from onemod.stage.data_stage.preprocessing_stage import PreprocessingStage
+
+__all__ = [PreprocessingStage]
