+++ src/onemod/utils/subsets.py
